requirements.txt
dill
dwave-neal
dwave_networkx
gurobipy
hyperopt
matplotlib
numpy
pandas
scipy